fix for travis

+++ db/migrate/20131216101839_remove_by_ids.rb
@@ -1,5 +1,7 @@
 class RemoveByIds < ActiveRecord::Migration
   def change
-    remove_index :indicator_data, { name: "by_ids" }
+    if index_exists?(:indicator_data, { name: "by_ids" })
+      remove_index :indicator_data, { name: "by_ids" }
+    end
   end
 end
